+++ ext_emconf.php
@@ -12,8 +12,8 @@ $EM_CONF[$_EXTKEY] = [
     'version' => '1.1.0-dev',
     'constraints' => [
         'depends' => [
-            'schema' => '1.7.0-1.99.99',
-            'typo3' => '9.5.0-11.5.99',
+            'schema' => '1.7.0-2.99.99',
+            'typo3' => '9.5.16-11.5.99',
         ],
         'conflicts' => [],
         'suggests' => [],