+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Allow usage with EXT:schema v2
+
 ## [1.0.1] - 2021-04-08
 
 ### Fixed
@@ -16,6 +19,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release with schema definition version 12.0
 
 
-[Unreleased]: https://github.com/brotkrueml/schema-pending/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/brotkrueml/schema-pending/compare/v1.0.1...HEAD
 [1.0.1]: https://github.com/brotkrueml/schema-pending/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/brotkrueml/schema-pending/releases/tag/v1.0.0